+++ src/stream.rs
@@ -5,8 +5,6 @@ use fut::ActorFuture;
 use actor::{Actor, ActorState, ActorContext, AsyncContext, SpawnHandle};
 
 /// Stream handler
-///
-/// `StreamHandler` is an extension of a `Handler` with stream specific methods.
 #[allow(unused_variables)]
 pub trait StreamHandler<I, E> where Self: Actor
 {